:recycle: Refactor balance views
